requirements.txt
# transformers==4.51.3 
transformers==4.57.1
# 使用qwen-omni模型时，transformers==4.54.0
# 使用qwen3-vl, qwen3-next模型时，transformers==4.57.1
datasets==2.21.0+openlm
# jemalloc coredump https://github.com/apache/arrow/issues/44342 
pyarrow==15.0.0
# 否则和transformer冲突：https://github.com/instructlab/instructlab/issues/3456
accelerate==1.7.0
peft==0.12.0
trl==0.9.6
tyro==0.8.14
scipy
einops
sentencepiece
protobuf
jieba
rouge-chinese
nltk==3.8.1
uvicorn
pydantic
pynvml>=12.0.0
pillow>=10.1.0
fastapi
sse-starlette
matplotlib
# deepspeed==0.16.2
transformers_stream_generator
auto-gptq
optimum
tiktoken
numpy==1.26.4
openlm-hub>=0.0.64
pytest
ml_tracker
librosa
av
timm
liger_kernel
qwen-omni-utils[decord]
faiss-gpu
